+++ scanner/PhotoAlbum.py
@@ -81,6 +81,7 @@ class Album(object):
 			self.password_identifiers_set = set()
 			self.passwords_marker_mtime = None
 			self.album_ini_mtime = None
+			self._date = datetime(1900, 1, 1)
 
 			if (
 				Options.config['subdir_method'] in ("md5", "folder") and
@@ -174,14 +175,15 @@ class Album(object):
 
 	@property
 	def date(self):
-		self.sort_subalbums_and_media()
+		self.sort_media_by_date()
+		self.sort_subalbum_by_date()
 		if len(self.media_list) == 0 and len(self.subalbums_list) == 0:
 			return datetime(1900, 1, 1)
 		elif len(self.media_list) == 0:
-			return self.subalbums_list[-1].date
+			return self.subalbums_list[-1]._date
 		elif len(self.subalbums_list) == 0:
 			return self.media_list[-1].date
-		return max(self.media_list[-1].date, self.subalbums_list[-1].date)
+		return max(self.media_list[-1].date, self.subalbums_list[-1]._date)
 
 	@property
 	def date_string(self):
@@ -198,7 +200,17 @@ class Album(object):
 
 	def __lt__(self, other):
 		try:
-			return self.date < other.date
+			if Options.config['default_album_name_sort'] or self.date == other.date:
+				if Options.config['default_album_reverse_sort']:
+					return self.name > other.name
+				else:
+					return self.name < other.name
+			else:
+				if Options.config['default_album_reverse_sort']:
+					return self.date > other.date
+				else:
+					return self.date < other.date
+			# return self.date < other.date
 		except TypeError:
 			return False
 
@@ -251,6 +263,16 @@ class Album(object):
 	def add_subalbum(self, album):
 		self.subalbums_list.append(album)
 		self.subalbums_list_is_sorted = False
+
+	def sort_subalbum_by_date(self):
+		self.subalbums_list.sort(key=lambda _subalbum: _subalbum._date)
+		if not Options.config['default_album_name_sort'] and not Options.config['default_album_reverse_sort']:
+			self.subalbums_list_is_sorted = False
+
+	def sort_media_by_date(self):
+		self.media_list.sort(key=lambda _media: _media.date)
+		if not Options.config['default_media_name_sort'] and not Options.config['default_media_reverse_sort']:
+			self.media_list_is_sorted = False
 
 	def sort_subalbums_and_media(self):
 		if not self.media_list_is_sorted:
@@ -2558,17 +2580,23 @@ class Media(object):
 			return ""
 
 	def __eq__(self, other):
-		return self.path == other.path
+		return self.album_path == other.album_path
 
 	# def __ne__(self, other):
 	# 	return not self.__eq__(other)
 
 	def __lt__(self, other):
 		try:
-			if self.date == other.date:
-				return self.name < other.name
+			if Options.config['default_media_name_sort'] or self.date == other.date:
+				if Options.config['default_media_reverse_sort']:
+					return self.name > other.name
+				else:
+					return self.name < other.name
 			else:
-				return self.date < other.date
+				if Options.config['default_media_reverse_sort']:
+					return self.date > other.date
+				else:
+					return self.date < other.date
 		except TypeError:
 			return False