+++ web/js/034-libphotofloat.js
@@ -558,14 +558,11 @@
 									function getSingleProtectedCacheBaseWithExternalMediaAndPositions_resolved() {
 										// ok, we got what we were looking for: numsProtectedMediaInSubTree property has been added by getSingleProtectedCacheBaseWithExternalMediaAndPositions()
 
-										if (self.hasOwnProperty("media")) {
-											self.media.sortByDate();
-											self.mediaNameSort = false;
-											self.mediaReverseSort = false;
-										}
-										util.sortByDate(self.subalbums);
-										self.albumNameSort = false;
-										self.albumReverseSort = false;
+										delete self.mediaNameSort;
+										delete self.mediaReverseSort;
+										delete self.albumNameSort;
+										delete self.albumReverseSort;
+										self.initializeSortPropertiesAndCookies();
 										self.sortAlbumsMedia();
 
 										resolve_getNextProtectedDirectory();
@@ -789,45 +786,16 @@
 								function() {
 									// execution arrives here when all the protected json has been loaded and processed
 
-									if (util.isSearchRootCacheBase(self.cacheBase)) {
-										resolve_continueAddProtectedContent();
-									} else {
-										// before sorting, any subalbum must be converted to album, because their data are slightly different
-										let conversionPromises = [];
-										// self.subalbums.forEach(
-										// 	function convertSubalbum(subalbum, iSubalbum) {
-										// 		if (subalbum instanceof Subalbum) {
-										// 			let conversionPromise = new Promise(
-										// 				function(resolve_conversionPromise) {
-										// 					let toAlbumPromise = subalbum.toAlbum(null, {getMedia: false, getPositions: false});
-										// 					toAlbumPromise.then(
-										// 						function(album) {
-										// 							self.subalbums[iSubalbum] = album;
-										// 							resolve_conversionPromise();
-										// 						}
-										// 					);
-										// 				}
-										// 			);
-										// 			conversionPromises.push(conversionPromise);
-										// 		}
-										// 	}
-										// );
-										Promise.all(conversionPromises).then(
-											function() {
-												if (self.hasOwnProperty("media")) {
-													self.media.sortByDate();
-													self.mediaNameSort = false;
-													self.mediaReverseSort = false;
-												}
-												util.sortByDate(self.subalbums);
-												self.albumNameSort = false;
-												self.albumReverseSort = false;
-												self.sortAlbumsMedia();
-
-												resolve_continueAddProtectedContent();
-											}
-										);
+									if (! util.isSearchRootCacheBase(self.cacheBase)) {
+										delete self.mediaNameSort;
+										delete self.mediaReverseSort;
+										// }
+										delete self.albumNameSort;
+										delete self.albumReverseSort;
+										self.initializeSortPropertiesAndCookies();
+										self.sortAlbumsMedia();
 									}
+									resolve_continueAddProtectedContent();
 								}
 							);
 						}
@@ -1789,8 +1757,7 @@
 													let promises = [];
 													if (env.searchAlbum.media.length) {
 														// search albums need to conform to default behaviour of albums:
-														// json files have subalbums and media sorted by date not reversed
-														env.searchAlbum.media.sortByDate();
+														// json files have subalbums and media sorted according to options
 														env.searchAlbum.media.forEach(
 															function(singleMedia) {
 																let promise = new Promise(
@@ -1815,12 +1782,11 @@
 															}
 														);
 													}
-													env.searchAlbum.mediaNameSort = false;
-													env.searchAlbum.mediaReverseSort = false;
+													delete env.searchAlbum.mediaNameSort;
+													delete env.searchAlbum.mediaReverseSort;
 
 													if (env.searchAlbum.subalbums.length) {
-														// search albums need to conform to default behaviour of albums: json files have subalbums and media sorted by date not reversed
-														util.sortByDate(env.searchAlbum.subalbums);
+														// search albums need to conform to default behaviour of albums: json files have subalbums and media sorted according to options
 														env.searchAlbum.subalbums.forEach(
 															function(subalbum, iSubalbum) {
 																let promise = new Promise(
@@ -1830,17 +1796,6 @@
 																			function(subalbum) {
 																				env.searchAlbum.subalbums[iSubalbum] = subalbum;
 																				env.searchAlbum.subalbums[iSubalbum].generateAlbumCaptionForSearch();
-
-																				// // replace the converted subalbum in its parent's subalbums
-																				// let splittedCacheBase = subalbum.cacheBase.split(env.options.cache_folder_separator);
-																				// let parentCacheBase = splittedCacheBase.slice(0, splittedCacheBase.length - 1).join(env.options.cache_folder_separator);
-																				// var getAlbumPromise = PhotoFloat.getAlbum(parentCacheBase, null, {getMedia: false, getPositions: false});
-																				// getAlbumPromise.then(
-																				// 	function(parentAlbum) {
-																				// 		indexInParent = parentAlbum.subalbums.findIndex(parentSubalbum => parentSubalbum.isEqual(subalbum));
-																				// 	}
-																				// );
-
 																				resolve_subalbum();
 																			},
 																			function() {
@@ -1856,8 +1811,8 @@
 
 													Promise.all(promises).then(
 														function() {
-															env.searchAlbum.albumNameSort = false;
-															env.searchAlbum.albumReverseSort = false;
+															delete env.searchAlbum.albumNameSort;
+															delete env.searchAlbum.albumReverseSort;
 
 															if (env.searchAlbum.media.length || env.searchAlbum.subalbums.length) {
 																env.searchAlbum.initializeSortPropertiesAndCookies();